add navigation stacks component
